Add membership claiming process.
Claim membership with an invitation code. Link membership to the existing
identity, or create a new identity with either external account or email and
SMS verification.

- Add link to invitation emails.
- Add membership id to invite token.
- Add new verification token types for phone_number and email_address.
- Add some of the links to login and registration templates.
- Fix overlapping invite membership checking: ignore linked identities.

See merge request kamu/kamu!30

+++ base/models.py
@@ -8,9 +8,14 @@ import string
 from datetime import timedelta
 
 from django.conf import settings
+from django.core.validators import validate_email
 from django.db import models
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
+
+from identity.models import EmailAddress as EmailAddressType
+from identity.models import PhoneNumber as PhoneNumberType
+from role.models import Membership as MembershipType
 
 
 class TimeLimitError(Exception):
@@ -40,23 +45,46 @@ class TokenManager(models.Manager["Token"]):
         """
         return getattr(settings, "TOKEN_SECRET_KEY", settings.SECRET_KEY)
 
-    def _create_token(self, token_type, membership=None, email=None, phone=None, length=8) -> str:
+    def _create_token(
+        self,
+        token_type,
+        membership: MembershipType | None = None,
+        email_object: EmailAddressType | None = None,
+        phone_object: PhoneNumberType | None = None,
+        email_address: str = "",
+        phone_number: str = "",
+        length: int = 8,
+    ) -> str:
         """
         Create a new token. Removes existing tokens of the same type and linked object.
 
         Raises TimeLimitError if token creation is attempted too soon.
         Raises ValueError if linked object or token_type is missing.
         """
-        if (not membership and not email and not phone) or not token_type:
+        if (not any([membership, email_object, phone_object, email_address, phone_number])) or not token_type:
             raise ValueError("Missing attributes.")
         create_time_limit = getattr(settings, "TOKEN_TIME_LIMIT_NEW", 60)
         verification_tries = getattr(settings, "TOKEN_VERIFICATION_TRIES", 3)
         try:
-            token = self.get(membership=membership, email=email, phone=phone, token_type=token_type)
+            token = self.get(
+                membership=membership,
+                email_object=email_object,
+                phone_object=phone_object,
+                email_address=email_address,
+                phone_number=phone_number,
+                token_type=token_type,
+            )
         except Token.DoesNotExist:
             token = None
         except Token.MultipleObjectsReturned:
-            self.filter(membership=membership, email=email, phone=phone, token_type=token_type).delete()
+            self.filter(
+                membership=membership,
+                email_object=email_object,
+                phone_object=phone_object,
+                email_address=email_address,
+                phone_number=phone_number,
+                token_type=token_type,
+            ).delete()
             token = None
         if token:
             if token.created_at < timezone.now() - timedelta(seconds=create_time_limit):
@@ -64,48 +92,75 @@ class TokenManager(models.Manager["Token"]):
             else:
                 raise TimeLimitError
         secret = self._generate_secret(length=length)
-        token = Token()
+        token = Token(
+            tries_left=verification_tries,
+            token_type=token_type,
+            email_object=email_object,
+            phone_object=phone_object,
+            membership=membership,
+            email_address=email_address,
+            phone_number=phone_number,
+        )
         salt = self._generate_secret(length=64, alphabet=string.printable)
         token.hash = salt + (hashlib.sha256(f"{self._get_secret_key()}{salt}{secret}".encode()).hexdigest())
-        token.tries_left = verification_tries
-        token.token_type = token_type
-        token.email = email
-        token.phone = phone
-        token.membership = membership
+        if membership:
+            secret = f"{membership.pk}:{secret}"
         token.save()
         return secret
 
-    def create_email_verification_token(self, email) -> str:
+    def create_email_object_verification_token(self, email: EmailAddressType) -> str:
         """
         Create a new email verification token.
         """
-        return self._create_token("emailverification", email=email)
+        return self._create_token("emailobjectverif", email_object=email)
 
-    def create_sms_verification_token(self, phone) -> str:
+    def create_phone_object_verification_token(self, phone: PhoneNumberType) -> str:
         """
         Create a new SMS verification token.
         """
-        return self._create_token("smsverification", phone=phone)
+        return self._create_token("phoneobjectverif", phone_object=phone)
 
-    def create_email_login_token(self, email) -> str:
+    def create_email_login_token(self, email: EmailAddressType) -> str:
         """
         Create a new email login token.
         """
-        return self._create_token("emaillogin", email=email)
+        return self._create_token("emaillogin", email_object=email)
 
-    def create_sms_login_token(self, phone) -> str:
+    def create_phone_login_token(self, phone: PhoneNumberType) -> str:
         """
         Create a new SMS login token.
         """
-        return self._create_token("smslogin", phone=phone)
+        return self._create_token("phonelogin", phone_object=phone)
 
-    def create_invite_token(self, membership) -> str:
+    def create_email_address_verification_token(self, email_address: str) -> str:
+        """
+        Create a new email address verification token.
+        """
+        return self._create_token("emailaddrverif", email_address=email_address)
+
+    def create_phone_number_verification_token(self, phone_number: str) -> str:
+        """
+        Create a new phone number verification token.
+        """
+        return self._create_token("phonenumberverif", phone_number=phone_number)
+
+    def create_invite_token(self, membership: MembershipType) -> str:
         """
         Create a new invite token.
         """
         return self._create_token("invite", membership=membership, length=32)
 
-    def _validate_token(self, secret, token_type, membership=None, email=None, phone=None) -> bool:
+    def _validate_token(
+        self,
+        secret,
+        token_type,
+        membership=None,
+        email_object=None,
+        phone_object=None,
+        email_address="",
+        phone_number="",
+        remove_token=True,
+    ) -> bool:
         """
         Validates a token.
 
@@ -120,11 +175,25 @@ class TokenManager(models.Manager["Token"]):
         else:
             verification_time_limit = getattr(settings, "TOKEN_LIFETIME_INVITE", 30 * 24 * 60 * 60)
         try:
-            token = self.get(membership=membership, email=email, phone=phone, token_type=token_type)
+            token = self.get(
+                membership=membership,
+                email_object=email_object,
+                phone_object=phone_object,
+                email_address=email_address,
+                phone_number=phone_number,
+                token_type=token_type,
+            )
         except Token.DoesNotExist:
             return False
         except Token.MultipleObjectsReturned:
-            self.filter(membership=membership, email=email, phone=phone, token_type=token_type).delete()
+            self.filter(
+                membership=membership,
+                email_object=email_object,
+                phone_object=phone_object,
+                email_address=email_address,
+                phone_number=phone_number,
+                token_type=token_type,
+            ).delete()
             return False
         salt = token.hash[:64]
         secret_hash = token.hash[64:]
@@ -133,7 +202,8 @@ class TokenManager(models.Manager["Token"]):
             token.delete()
             return False
         if token.tries_left > 0 and secrets.compare_digest(secret_hash.encode(), token_hash.encode()):
-            token.delete()
+            if remove_token:
+                token.delete()
             return True
         token.tries_left -= 1
         token.save()
@@ -141,35 +211,47 @@ class TokenManager(models.Manager["Token"]):
             token.delete()
         return False
 
-    def validate_email_verification_token(self, secret, email) -> bool:
+    def validate_email_object_verification_token(self, secret: str, email: EmailAddressType) -> bool:
         """
         Validates a email verification token.
         """
-        return self._validate_token(secret, "emailverification", email=email)
+        return self._validate_token(secret, "emailobjectverif", email_object=email)
 
-    def validate_sms_verification_token(self, secret, phone) -> bool:
+    def validate_phone_object_verification_token(self, secret: str, phone: PhoneNumberType) -> bool:
         """
         Validates an SMS verification token.
         """
-        return self._validate_token(secret, "smsverification", phone=phone)
+        return self._validate_token(secret, "phoneobjectverif", phone_object=phone)
 
-    def validate_email_login_token(self, secret, email) -> bool:
+    def validate_email_login_token(self, secret: str, email: EmailAddressType) -> bool:
         """
         Validates a email login token.
         """
-        return self._validate_token(secret, "emaillogin", email=email)
+        return self._validate_token(secret, "emaillogin", email_object=email)
 
-    def validate_sms_login_token(self, secret, phone) -> bool:
+    def validate_phone_login_token(self, secret: str, phone: PhoneNumberType) -> bool:
         """
         Validates an SMS login token.
         """
-        return self._validate_token(secret, "smslogin", phone=phone)
+        return self._validate_token(secret, "phonelogin", phone_object=phone)
 
-    def validate_invite_token(self, secret, membership) -> bool:
+    def validate_email_address_verification_token(self, secret: str, email_address: str) -> bool:
+        """
+        Validates an email address verification token.
+        """
+        return self._validate_token(secret, "emailaddrverif", email_address=email_address)
+
+    def validate_phone_number_verification_token(self, secret: str, phone_number: str) -> bool:
+        """
+        Validates a phone number verification token.
+        """
+        return self._validate_token(secret, "phonenumberverif", phone_number=phone_number)
+
+    def validate_invite_token(self, secret: str, membership: MembershipType, remove_token: bool = True) -> bool:
         """
         Validates a invite login token.
         """
-        return self._validate_token(secret, "invite", membership=membership)
+        return self._validate_token(secret, "invite", membership=membership, remove_token=remove_token)
 
 
 class Token(models.Model):
@@ -178,14 +260,20 @@ class Token(models.Model):
     """
 
     membership = models.ForeignKey("role.Membership", null=True, on_delete=models.CASCADE)
-    email = models.ForeignKey("identity.EmailAddress", null=True, on_delete=models.CASCADE)
-    phone = models.ForeignKey("identity.PhoneNumber", null=True, on_delete=models.CASCADE)
+    email_object = models.ForeignKey("identity.EmailAddress", null=True, on_delete=models.CASCADE)
+    phone_object = models.ForeignKey("identity.PhoneNumber", null=True, on_delete=models.CASCADE)
+    phone_number = models.CharField(max_length=20, blank=True, verbose_name=_("Phone number"))
+    email_address = models.CharField(
+        max_length=320, blank=True, verbose_name=_("Email address"), validators=[validate_email]
+    )
 
     TOKEN_TYPE_CHOICES = (
-        ("emailverification", _("E-mail verification token")),
-        ("smsverification", _("SMS verification token")),
         ("emaillogin", _("E-mail login token")),
-        ("smslogin", _("SMS login token")),
+        ("phonelogin", _("SMS login token")),
+        ("emailobjectverif", _("E-mail object verification token")),
+        ("phoneobjectverif", _("Phone object verification token")),
+        ("emailaddrverif", _("E-mail address verification token")),
+        ("phonenumberverif", _("Phone number verification token")),
         ("invite", _("Invite token")),
     )
     token_type = models.CharField(max_length=17, choices=TOKEN_TYPE_CHOICES, verbose_name=_("Token type"))
